refactor(scenario-gen): move the declared-usage-site oracle into ground-truth (gpt-l9rp)
"Every place a generated workspace references an entity" follows from the
scenario by construction, which is exactly what `ground-truth.js` owns —
`scenarioDeclaredFieldPaths` and `scenarioFieldEdges` are the same class of
thing. R3 landed it in the LSP's test tree because the LSP was its first
consumer; leaving it there guaranteed the CLI would copy it the moment it grew a
references or rename property, and R4 is that consumer.

`declaredEntities`, `declaredUsageSites` and `entityKeyForRef` become
`scenarioDeclaredEntities`, `scenarioDeclaredUsageSites` and
`scenarioEntityKeyForRef`, matching the file's convention — at a call site the
prefix is what says the expectation came from the input rather than from the
toolchain. `USAGE_KIND` moves unrenamed. Eleven hand-written cases come with
them, including the import derivation the oracle restates from the renderer and
the `withheldImports` hole R1's mutators reach through.

Two things stay behind. `RESOLVABLE_USAGE_KINDS` lists what the LSP's definition
provider resolves *today* — a statement about that server, not ground truth — so
it sits beside the pinned tests for the kinds it excludes. And the adapter stays
in the LSP: the CLI's writes to disk and loads an import graph, this one indexes
in-memory documents, and sl-rw3e exists because those two scope duplicates
differently. Its module comment now says so.

`bareNamespacedWorkspaceArbitrary` moves to `workspace-arbitraries.js` but stays
out of `workspaceScenarioArbitrary`: the edge oracles canonicalise a bare ref as
file-scope, so folding this domain into the shared default would move every
consuming package's expectations. That reasoning now lives on the arbitrary
rather than in a test file no other package reads.

LSP suite unchanged at 317; scenario-gen 31 -> 42.
