+++ tooling/satsuma-scenario-gen/src/ground-truth.js
@@ -330,3 +330,311 @@ export function scenarioSchemaProjection(edges) {
   }
   return [...projected].sort();
 }
+
+// ── Declared entities, and every place a workspace uses them ───────────────
+//
+// The section above answers "what flows where". This one answers "what is
+// declared, and where is it named" — the ground truth behind find-references,
+// go-to-definition and rename. It lives here for the same reason
+// `scenarioFieldEdges` does: a generated workspace declares its entities and
+// names every reference to them outright, so the answer follows from the
+// scenario by construction and no consumer should re-derive it. The LSP was the
+// first consumer (Feature 46 R3) and the CLI is not far behind, which is what
+// `gpt-l9rp` moved this out of the LSP's test tree for.
+//
+// One rule *is* restated from `workspace-render.js`: which names a file's
+// derived `import` statements list. A scenario does not author its imports — the
+// renderer computes them from usage, so a generated workspace can never claim an
+// import graph its declarations contradict — but an imported name is a reference
+// site, and much of what the multi-file domain is for. The alternative is to
+// read the emitted `import` lines back out of rendered text, which would make an
+// oracle into a parser. Keeping the restatement in this package, beside the
+// renderer it mirrors, is what stops the two drifting.
+
+/**
+ * Usage kinds, spelled exactly as the workspace index's `ReferenceEntry.context`
+ * so that an expected site and an observed one compare as plain strings.
+ */
+export const USAGE_KIND = Object.freeze({
+  /** An entity named in a mapping's `source { }` list. */
+  source: "source",
+  /** An entity named in a mapping's `target { }` list. */
+  target: "target",
+  /** A fragment spread into a schema or fragment body. */
+  spread: "spread",
+  /** A name listed in an `import { }` declaration. */
+  import: "import",
+  /** A schema named by a metric's `source` metadata token. */
+  metricSource: "metric_source",
+  /** The schema prefix of a qualified arrow path, e.g. the `s0` of `s0.field`. */
+  arrow: "arrow",
+});
+
+/**
+ * An entity a generated workspace declares.
+ *
+ * @typedef {{
+ *   key: string,
+ *   name: string,
+ *   namespace: string | null,
+ *   keyword: "schema" | "fragment" | "mapping",
+ *   file: string,
+ * }} ScenarioDeclaredEntity
+ *
+ * `key` is the workspace index's definition key — `name` at file scope,
+ * `ns::name` inside a namespace — and therefore the key a reference query must
+ * arrive at. `name` is the block label as authored, which is what appears at the
+ * declaration site and inside a namespace is *not* the key. `keyword` is the
+ * declaration keyword, which is how the declaration is located in rendered text.
+ * `file` is the workspace-relative path of the declaring file.
+ */
+
+/**
+ * Every entity a workspace declares, in declaration order.
+ *
+ * Mappings are included even though nothing may reference them: "no references"
+ * is an assertion worth making, and it is the half of a reference property that
+ * catches an invented reference rather than a missing one.
+ *
+ * @param {import("./workspace-model.js").ScenarioWorkspace} workspace
+ * @returns {ScenarioDeclaredEntity[]}
+ */
+export function scenarioDeclaredEntities(workspace) {
+  const entities = [];
+  for (const file of workspace.files) {
+    for (const fragment of file.fragments) {
+      entities.push({
+        key: fragment.name,
+        name: fragment.name,
+        namespace: null,
+        keyword: "fragment",
+        file: file.path,
+      });
+    }
+    for (const schema of file.schemas) {
+      entities.push({
+        key: declarationKey(schema.name, schema.namespace),
+        name: schema.name,
+        namespace: schema.namespace,
+        keyword: "schema",
+        file: file.path,
+      });
+    }
+    for (const mapping of file.mappings) {
+      entities.push({
+        key: declarationKey(mapping.name, mapping.namespace),
+        name: mapping.name,
+        namespace: mapping.namespace,
+        keyword: "mapping",
+        file: file.path,
+      });
+    }
+  }
+  return entities;
+}
+
+/**
+ * Every usage site the workspace declares, grouped by the entity it references.
+ *
+ * Every declared entity gets an entry, empty when nothing references it. Sites
+ * are a **multiset**, not a set: one file legitimately references the same entity
+ * from two mappings, and collapsing that to one site would hide a reference the
+ * toolchain dropped.
+ *
+ * Two things are deliberately *not* sites. A `namespace` block is not one: the
+ * scenario model has no namespace declaration to name, and a qualified reference
+ * is filed under the whole `ns::name` key. Nor is the schema an NL `@ref`
+ * mentions: `@raw.field_1` is a reference to the *field* `raw.field_1`, which is
+ * how the index files it. The `@ref` still matters here, because a mention of
+ * another file's schema is one of the things that forces an `import`, which is a
+ * site.
+ *
+ * Throws when the scenario references something it never declares — that is a
+ * malformed scenario, not a toolchain failure, and failing loudly beats silently
+ * dropping the site.
+ *
+ * @param {import("./workspace-model.js").ScenarioWorkspace} workspace
+ * @returns {Map<string, Array<{ file: string, kind: string }>>}
+ */
+export function scenarioDeclaredUsageSites(workspace) {
+  const entities = scenarioDeclaredEntities(workspace);
+  const keys = new Set(entities.map((entity) => entity.key));
+  const sites = new Map(entities.map((entity) => [entity.key, []]));
+
+  /** Record one site, resolving the authored spelling to the entity it names. */
+  const add = (ref, namespace, file, kind) => {
+    const key = scenarioEntityKeyForRef(ref, namespace, keys);
+    const forEntity = sites.get(key);
+    if (!forEntity) {
+      throw new Error(`scenario error: '${ref}' in ${file} names no declared entity`);
+    }
+    forEntity.push({ file, kind });
+  };
+
+  for (const file of workspace.files) {
+    for (const schema of file.schemas) {
+      for (const ref of schema.metricSources ?? []) {
+        add(ref, schema.namespace, file.path, USAGE_KIND.metricSource);
+      }
+      for (const ref of schema.spreads ?? []) {
+        add(ref, schema.namespace, file.path, USAGE_KIND.spread);
+      }
+    }
+    for (const mapping of file.mappings) {
+      for (const ref of mapping.sources) add(ref, mapping.namespace, file.path, USAGE_KIND.source);
+      for (const ref of mapping.targets) add(ref, mapping.namespace, file.path, USAGE_KIND.target);
+      for (const ref of qualifiedArrowSchemas(mapping)) {
+        add(ref, mapping.namespace, file.path, USAGE_KIND.arrow);
+      }
+    }
+    for (const ref of importedRefs(file, workspace)) {
+      // Import declarations are top-level, so they are authored in no namespace.
+      add(ref, null, file.path, USAGE_KIND.import);
+    }
+  }
+
+  return sites;
+}
+
+/**
+ * The entity an authored reference names, given the namespace it was written in.
+ *
+ * Satsuma's scoping rule (spec §5.3): a name written with `::` names that entity
+ * outright; a bare name binds to the enclosing namespace's declaration when there
+ * is one, and to the file-scope declaration otherwise. This is a *language* rule,
+ * so stating it here re-derives nothing the toolchain decides — but it is also
+ * the rule the workspace index's `resolveReferenceKey` implements, and getting it
+ * right in one place and wrong in the other is `sl-p256`.
+ *
+ * Note this is a different question from {@link canonicalEntityRef}, which maps a
+ * *canonical* ref onto the `[ns]::name` form the edge oracles compare with. That
+ * one reads a bare ref as file-scope unconditionally, which is why
+ * `bareNamespacedWorkspaceArbitrary` is not part of the shared default domain.
+ *
+ * @param {string} ref authored spelling, e.g. `"s0"` or `"ns_a::s0"`
+ * @param {string | null} namespace the namespace block the reference sits in
+ * @param {Set<string>} declaredKeys every entity key the workspace declares
+ */
+export function scenarioEntityKeyForRef(ref, namespace, declaredKeys) {
+  if (ref.includes("::")) return ref;
+  if (namespace && declaredKeys.has(`${namespace}::${ref}`)) return `${namespace}::${ref}`;
+  return ref;
+}
+
+/** The index key for a declaration: `ns::name` inside a namespace, `name` outside. */
+function declarationKey(name, namespace) {
+  return namespace ? `${namespace}::${name}` : name;
+}
+
+/**
+ * The schemas named by *qualified* arrow paths in a mapping.
+ *
+ * An arrow endpoint is `{ schema, path }` in the model, and the renderer decides
+ * how it is spelled: bare when its side of the mapping has exactly that one
+ * schema, `schema.path` when the side has several, and `.suffix` inside a
+ * container block. Only the middle form writes the schema down, and the index
+ * files the first segment of a path as a reference — which is what makes renaming
+ * a schema rewrite `s0.field` correctly.
+ *
+ * A *namespaced* schema on a multi-schema side would be spelled `ns::schema.path`
+ * and filed by the index under the bare `schema`. No generated domain produces
+ * that shape today; if one ever does, this oracle will correctly report the site
+ * as missing rather than quietly agreeing with whatever the index did.
+ */
+function qualifiedArrowSchemas(mapping) {
+  const schemas = [];
+
+  const visit = (arrows, insideContainer) => {
+    for (const arrow of arrows) {
+      // Inside a container block every path is relative (`.suffix`), so it names
+      // no schema at all.
+      if (!insideContainer) {
+        for (const endpoint of sourceEndpointsOf(arrow)) {
+          if (namesItsSchema(endpoint, mapping.sources)) schemas.push(endpoint.schema);
+        }
+        if (namesItsSchema(arrow.target, mapping.targets)) schemas.push(arrow.target.schema);
+      }
+      if (arrow.children) visit(arrow.children, true);
+    }
+  };
+
+  visit(mapping.arrows, false);
+  return schemas;
+}
+
+/** The source endpoints of any arrow kind: several for `map`, one for a container header, none for `computed`. */
+function sourceEndpointsOf(arrow) {
+  if (arrow.sources) return arrow.sources;
+  return arrow.source ? [arrow.source] : [];
+}
+
+/** Mirrors the renderer's `authoredEndpoint`: a side with one matching schema writes the path bare. */
+function namesItsSchema(endpoint, sideSchemas) {
+  return !(sideSchemas.length === 1 && sideSchemas[0] === endpoint.schema);
+}
+
+/**
+ * The names a file's derived `import` statements list.
+ *
+ * Restates `workspace-render.js`'s `renderImports`: everything the file
+ * references but does not declare, that some other file declares, once per
+ * entity, minus anything the scenario deliberately withholds. See this section's
+ * header for why the rule is restated rather than read back out of rendered text.
+ */
+function importedRefs(file, workspace) {
+  const own = declaredRefsOf(file);
+  // `withheldImports` is the renderer's one deliberate hole in the derivation — a
+  // defect mutator's way of reaching ADR-022's import-scope check. A valid
+  // scenario leaves it empty, but honouring it here keeps a withheld import
+  // reported as the missing *import statement* it is, rather than as a
+  // find-references failure against a site the file never wrote.
+  const withheld = new Set(file.withheldImports ?? []);
+  const wanted = [...referencedRefsOf(file)].filter((ref) => !own.has(ref) && !withheld.has(ref));
+
+  const imported = [];
+  for (const other of workspace.files) {
+    if (other.path === file.path) continue;
+    const declared = declaredRefsOf(other);
+    for (const ref of wanted) {
+      if (declared.has(ref)) imported.push(ref);
+    }
+  }
+  return imported;
+}
+
+/**
+ * Every entity reference a file makes, as authored.
+ *
+ * Wider than the list of *sites*: an arrow path written bare still needs its
+ * schema imported, and so does the schema an NL `@ref` mentions. This is the list
+ * the import derivation is computed from, not the list of reference sites.
+ */
+function referencedRefsOf(file) {
+  const refs = new Set();
+  for (const schema of file.schemas) {
+    for (const ref of schema.metricSources ?? []) refs.add(ref);
+    for (const ref of schema.spreads ?? []) refs.add(ref);
+  }
+  for (const mapping of file.mappings) {
+    for (const ref of [...mapping.sources, ...mapping.targets]) refs.add(ref);
+    for (const arrow of flattenArrows(mapping.arrows)) {
+      for (const endpoint of sourceEndpointsOf(arrow)) refs.add(endpoint.schema);
+      refs.add(arrow.target.schema);
+      for (const endpoint of arrow.transform?.refs ?? []) refs.add(endpoint.schema);
+    }
+  }
+  return refs;
+}
+
+/**
+ * Every entity a file declares, spelled the way a reference to it is written.
+ *
+ * Mappings are absent on purpose: the renderer's import derivation does not
+ * consider them either, because nothing may reference a mapping.
+ */
+function declaredRefsOf(file) {
+  return new Set([
+    ...file.fragments.map((fragment) => fragment.name),
+    ...file.schemas.map((schema) => declarationKey(schema.name, schema.namespace)),
+  ]);
+}
